Added another fix for isCurrent.
git-svn-id: file:///Users/lightster/backup-8.31.2013_20-06-44_litesign/homedir/svn/Workspaces/Dev@115 3baa19a1-490e-4327-b14f-f43bf3c759de

+++ nascar/src/com/lightdatasys/nascar/live/gui/cell/ResultCell.java
@@ -134,7 +134,7 @@ public class ResultCell extends Cell
 		}
 		else if(mode == Mode.LAST_LAP_SPEED)
 		{
-			return String.format("%d", (int)result.getLastLapSpeed());
+			return String.format("%.2f", result.getLastLapSpeed());
 		}
 		else if(mode == Mode.SPEED)
 		{
